remove the unnecessary build target
