+++ packages/database/index.ts
@@ -35,7 +35,7 @@ const ServerValue = Database.ServerValue;
 export function registerDatabase(instance: FirebaseNamespace) {
   
   // set SDK_VERSION
-  setSDKVersion(firebase.SDK_VERSION);
+  setSDKVersion(instance.SDK_VERSION);
   
   // Register the Database Service with the 'firebase' namespace.
   const namespace = (instance as _FirebaseNamespace).INTERNAL.registerService(
